pyproject.toml
[tool.poetry]
name = "web-project"
version = "0.1.0"
description = ""
authors = ["Vlad Vladov <vlad.vladov@alemira.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.4"
psycopg = {version = "^3.1.10", extras = ["binary"]}
pillow = "^10.0.0"
djangorestframework = "^3.14.0"
django-filter = "^23.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
